refactor: move _check_data_dir to cli.py, where it is used the most
